Working. Was able to recreate classic images.

+++ src/logic/mandelbrot_iteration.rs
@@ -1,34 +1,6 @@
 use num::complex::Complex;
 
-// export function mapRange(n: number, sourceMin: number, sourceMax: number, targetMin: number, targetMax: number): number {
-// const sourceRatio = (n - sourceMin) / (sourceMax - sourceMin);
-//
-// return (sourceRatio * (targetMax - targetMin)) + targetMin;
-// }
-
-// import * as math from 'mathjs';
-// import {Complex} from 'mathjs';
-//
-// // Port of my old C code
-// // https://codereview.stackexchange.com/questions/217574/ascii-mandelbrot-set-image-producer
-//
-//
-// export function testPoint(initial: Complex, maxIterations: number, infinityLimit: number = 2): number {
-//   let current = initial;
-//
-//   let i;
-//   for (i = 0; i < maxIterations; i++) {
-//     if (isUnderLimit(current, infinityLimit)) {
-//       current = mandelbrotIteration(initial, current);
-//     } else {
-//       break;
-//     }
-//   }
-//
-//   return i;
-// }
-
-pub type ComplexPoint = Complex<i64>;
+pub type ComplexPoint = Complex<f64>;
 
 fn mandelbrot_iteration(initial: ComplexPoint, current: ComplexPoint) -> ComplexPoint {
     let current_sq = current * current;
@@ -42,10 +14,9 @@ fn is_under_limit(current: ComplexPoint, infinity_limit: u32) -> bool {
     let real_sq = current.re * current.re;
     let imag_sq = current.im * current.im;
 
-    return (real_sq + imag_sq) <= (infinity_limit * infinity_limit) as i64;
+    return (real_sq + imag_sq) <= (infinity_limit * infinity_limit) as f64;
 }
 
-// TODO: u32 for max_iterations and the return can be generic technically
 pub fn test_point(initial: ComplexPoint, max_iterations: u32, infinity_limit: u32) -> u32 {
     let mut current = initial;
